# auslib/web/admin/swagger/api.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "Balrog's Admin API"
  termsOfService: "https://wiki.mozilla.org/Balrog"
  license:
    name: "Mozilla Public License 2.0"
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  contact:
    name: "Mozilla Release Engineering"
    url: "https://github.com/mozilla/balrog"
    email: "release-engineering@lists.mozilla.org"
  description: >
    **Balrog** is the software that runs the server side component of the update
    system used by Firefox and other Mozilla products. It is the successor to *AUS*
    (Application Update Service), which did not scale to our current needs nor allow
    us to adapt to more recent business requirements. Balrog helps us ship updates faster
    and with much more flexibility than we’ve had in the past.
security:
  - basic: []
securityDefinitions:
  basic:
    type: basic
    description: HTTP basic access authentication

# host and scheme can be omitted for a more dynamic association.
# In this case, the host and scheme used to serve the API documentation
# will be used for API calls. For example, if Swagger UI-based documentation is
# hosted at https://api.example.com/apidocs/index.html, “try it out” API calls
# will be directed to https://api.example.com.

#schemes:
#  - http
#  - https
# basePath: "/api"
# host: "localhost:8080"
# Swagger-UI bug : fix TODO week 13
...
parameters:
...
  changedByParam:
    name: changed_by
    in: query
    description: The username who made the requested changes.
    type: string
    x-nullable: true
    required: false
  changeIdParam:
    name: change_id
    in: query
    description: The change_id value to filter hsitory by.
    type: integer
    x-nullable: true
    required: false
  historyDataVersionParam:
    name: data_version
    in: query
    description: data_version field value to filter history by.
    type: integer
    x-nullable: true
    required: false
  timestampParam:
    name: timestamp
    in: query
    description: timestamp at which the requested change was made.
    type: integer
    x-nullable: true
    required: false
  timestampFromParam:
    name: timestamp_from
    in: query
    description: timestamp from/above which to filter history.
    type: integer
    x-nullable: true
    required: false
  timestampToParam:
    name: timestamp_to
    in: query
    description: timestamp to/below which history is filtered.
    type: integer
    x-nullable: true
    required: false
  channelParam:
    name: channel
    in: query
    description: channel stirng for filtering history.
    type: string
    x-nullable: true
    required: false
responses:
...
  newObjectDataVersion:
    description: "A new object (permission/role/release/signoff) is successfully created and its data version value is always 1"
    schema:
      type: object
      required:
        - new_data_version
      properties:
        new_data_version:
          description: "data version of the new object (permission/role/release/signoff) assigned. Value is always 1"
          type: integer
          format: int32
          minimum: 1
          maximum: 1
          example: 1

  # explicit Response for APIs utilizing changeRelease() method to serve requests
  # Also used in other APIs where data_version might not change after operation.
  # Hence no minLimit: 1 , no maxLimit on new_data_version.
  # Example: when attempting assign an already existing role to a user.
...
paths:
...
  /rules/history:
    get:
      summary: Returns a list of all histories.
      description: >
        Returns a list of rules history.
      tags:
        - "History Histories Rules"
      operationId: auslib.web.common.history_all.rules_history
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of rules history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - $ref: '#/parameters/channelParam'
      responses:
        '200':
          description: Successfully fetched rules history
        '400':
          $ref: '#/responses/invalidFormData'
  /releases/history:
    get:
      summary: Returns a list of releases history.
      description: >
        Returns a list of releases history.
      tags:
        - "History Histories Releases"
      operationId: auslib.web.common.history_all.releases_history
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of releases history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - $ref: '#/parameters/channelParam'
      responses:
        '200':
          description: Successfully fetched histories
        '400':
          $ref: '#/responses/invalidFormData'
  /permissions/history:
    get:
      summary: Returns a list of permissions history.
      description: >
        Returns a list of permissions history.
      tags:
        - "History Histories Permissions"
      operationId: auslib.web.common.history_all.permissions_history
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of rules, releases and permissions histories"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - $ref: '#/parameters/channelParam'
      responses:
        '200':
          description: Successfully fetched permissions history
        '400':
          $ref: '#/responses/invalidFormData'
  /product_required_signoffs/history:
    get:
      summary: Returns a list of product required signoff history.
      description: >
        Returns product required signoff history.
      tags:
        - "History Signoff Product"
      operationId: auslib.web.common.history_all.product_required_signoffs_history
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of product required signoff history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - name: role
          in: query
          description: The role(s) required for product signoff
          type: string
          x-nullable: true
          required: false
      responses:
        '200':
          description: Successfully fetched product required signoffs history
        '400':
          $ref: '#/responses/invalidFormData'
  /permissions_required_signoffs/history:
    get:
      summary: Returns a list of permissions required signoffs history.
      description: >
        Returns permissions required signoff history.
      tags:
        - "History Permissions Signoff"
      operationId: auslib.web.common.history_all.permissions_required_signoffs_history
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of permissions required signoff history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - name: role
          in: query
          description: The role(s) required for permission signoff
          type: string
          x-nullable: true
          required: false
      responses:
        '200':
          description: Successfully fetched permissions required signoff history
        '400':
          $ref: '#/responses/invalidFormData'
...
  /releases:
    get:
      summary: "Returns a JSON Object containing metadata about Releases in Balrog's database"
      description: >
        Returns a JSON object containing metadata about Releases in Balrog's database. Due to its size, the actual Release 'blob' is never returned from this endpoint. There are a few query arguments that affect its response. If no arguments are provided, it returns all information about all of the Releases in the database. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#releases)
      tags:
        - Releases
      operationId: auslib.web.admin.views.mapper.release_get
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#releases"
        description: "Returns a JSON Object containing metadata about Releases in Balrog's database."
      parameters:
        - $ref: '#/parameters/productParam'
        - name: name_prefix
          in: query
          description: "If 'name_prefix' is passed, only Releases whose name starts with the given prefix will be returned."
          type: string
          x-nullable: true
          required: false
        - name: names_only
          in: query
          description: "If 'names_only' is set to true, the response changes format and provides a list of all of the Release names in the database"
          type: string
          x-nullable: true
          required: false
      responses:
        '200':
          description: Returns metadata about Releases
          schema:
            type: object
            # since there are two exclusive and independent data formats i.e. releases and names and
            # oneOf isn't supported in swagger 2.0 hence cannot have 'required'
            # field in schema
            properties:
              releases:
                description: list of release objects
                type: array
                minItems: 0
                items:
                  allOf:
                    - $ref: '#/definitions/ReleaseGET'
                    - properties:
                        required_signoffs:
                          description: Object mapping roles to the number of signoffs required to change the release
                          type: object

              names:
                description: List of release names
                type: array
                minItems: 0
                uniqueItems: true
                items:
                  type: string
                  maxLength: 100
          examples:
            application/json:
              releases:
                - data_version: 2
                  name: XP-Vista-Desupport
                  product: Firefox
                  read_only: false
                  rule_ids: [468, 498]

    post:
      summary: Creates a new Release
      description: >
        Creates a new Release with the provided values. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id9)
      tags:
        - Releases
      operationId: auslib.web.admin.views.mapper.release_post
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id9"
        description: "Creates a new Release in Balrog's DB."
      parameters:
        - name: release_body
          in: body
          description: New Release object data
          required: true
          schema:
            $ref: '#/definitions/ReleaseData'
      responses:
        '201':
          $ref: '#/responses/newObjectDataVersion'
        '400':
          $ref: '#/responses/invalidFormData'
...
  /required_signoffs/product/history:
    get:
      summary: Returns a list of product required signoff history.
      description: >
        Returns product required signoff history.
      tags:
        - "History Signoff Product"
      operationId: auslib.web.common.history_all.get_product_required_signoffs_history
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of product required signoff history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
        - name: role
          in: query
          description: The role(s) required for product signoff
          type: string
          x-nullable: true
          required: false
      responses:
        '200':
          description: Successfully fetched product required signoff history
        '400':
          $ref: '#/responses/invalidFormData'
...
  /rules:
    get:
      summary: Returns filtered list of Rules from Balrog's DB
      description: >
        Returns filtered list Rules in Balrog’s database inside of a JSON Object. Returns all rules if 'product' query parameter isn't passed. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id1)
      tags:
        - Rules
      operationId: auslib.web.common.rules.get_rules
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id1"
        description: "Returns filtered list of Rules from Balrog's DB."
      parameters:
        - $ref: '#/parameters/productParam'

      responses:
        '200':
          description: Get json object of list all Rules with their count
          schema:
            $ref: '#/definitions/RulesListGET'

    post:
      summary: creates a new Rule in Balrog's DB
      description: >
        Creates a new Rule in Balrog's database. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#post)
      tags:
        - Rules
      operationId: auslib.web.admin.views.mapper.rules_post
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id1"
        description: "creates a new Rule in Balrog's DB."
      parameters:
        - name: rule
          in: body
          description: New Rule object data
          required: true
          schema:
            allOf:
              - $ref: '#/definitions/CSRFModel'
              - $ref: '#/definitions/UpdateType'
              - $ref: '#/definitions/RulesBase'
            #If you change the required list order or add extra fields, make sure to modify this endpoint's test cases plus code for handling /scheduled_changes/rules/  POST operation endpoint.
            required:
              - update_type
              - backgroundRate
              - priority
      responses:
        '200':
          description: Successfully created a new rule
          schema:
            title: successfully created a new rule
            type: integer
            example: 514
          examples:
            text/html: 514
        '400':
          $ref: '#/responses/invalidFormData'
...
definitions:
  CSRFModel:
    title: CSRF Token
    description: CSRF Token in POST/PUT or UPDATE requests
    type: object
    required:
      - csrf_token
    properties:
      csrf_token:
        type: string
        example: "1491342563##c4e6fef0b978e6c89af9ff1015e67b9ca7c45d14"
...
  ReleaseData:
    title: Release form values plus blob
    description: "Release Base Model + release blob"
    type: object
    allOf:
      - $ref: "#/definitions/CSRFModel"
      - $ref: '#/definitions/DataVersionModel'
      - $ref: '#/definitions/ProductModel'
      - $ref: '#/definitions/ReleaseName'
      - type: object
        properties:
          blob:
            type: string
            example: '{"hashFunction": "sha512", "name": "CDM-979", "product": "Firefox", "schema_version": 1000, "vendors": { "gmp-eme-adobe": { "platforms": {"WINNT_x86_64-msvc-x64": {"alias": "WINNT_x86_64-msvc"}},"version": "15"}}}'
            format: JSONStringField
            description: json blob containing information about the release
    required:
      - name
      - product
      - blob
...